More stable code

+++ cocos2dx/label_nodes/CCLabelBMFont.cpp
@@ -464,7 +464,7 @@ namespace cocos2d{
         }
 
         totalHeight = m_pConfiguration->m_uCommonHeight * quantityOfLines;
-        nextFontPositionY = -(m_pConfiguration->m_uCommonHeight - m_pConfiguration->m_uCommonHeight * quantityOfLines);
+        nextFontPositionY = -((long)(m_pConfiguration->m_uCommonHeight) - (long)(m_pConfiguration->m_uCommonHeight * quantityOfLines));
 
 		for (unsigned int i= 0; i < stringLen; i++)
 		{
